Loop in analysis solved.
Fixed a bug in expression comparison preventing result update from
reaching fixed point.
Now field update procedure must be fixed.

+++ src/com/laneve/asp/ASMAnalysis/asmTypes/expressions/ConstExpression.java
@@ -35,7 +35,7 @@ protected Long constExp;
 	@Override
 	public boolean equalValue(AnValue iExpression) {
 		if (!(iExpression instanceof ConstExpression)) return false;
-		return constExp == ((ConstExpression)iExpression).constExp;
+		return constExp.intValue() == ((ConstExpression)iExpression).constExp.intValue();
 	}
 
 }
